Remove unused debug code
Spelling fix
Try detach multiple times because something in MacOS can sometimes hold the directory
VMS-8527: System notification 'Storage is not configured' doesn't appear
VMS-8571: Add PUT and DELETE methods to HTTP request as an action
VMS-8597: Correctly update 'VMAX requires license' message on cameras on license activation
VMS-8620: Support hardware key windows executables signing
VMS-8631: Checkbox 'Launch videowall when Windows starts' should be set by default
VMS-8632: Disable screen recording for user without 'Export archive' permission
vms-8525: Merge with vms_3.2_dev
vms-8525: Merge with x_vms_8525_root_tool_refactoring
vms-8525: Refactor common root_tool code to separate library
win server startup fix
windows compile fix

<==== Mercurial data ====>
Branch: vms_3.2

+++ macsetup/client-dmg/maven/filter-resources/build_dmg.sh
@@ -88,7 +88,18 @@ mkdir -p $TMP
 
 hdiutil attach "raw-$DMG_FILE" -mountpoint "$TMP"
 SetFile -a C "$TMP"
-hdiutil detach "$TMP"
+
+# Try detach multiple times because something in MacOS can sometimes hold the directory.
+DETACHED=0
+for i in {1..20}
+do
+    hdiutil detach "$TMP" && DETACHED=1 && break
+    echo "Cannot detach $TMP" >&2
+    (set -x && lsof "$TMP")
+    sleep 1
+done
+[ $DETACHED -ne 1 ] && exit 1
+
 rm -rf "$TMP"
 rm -f "$DMG_FILE"
 hdiutil convert "raw-$DMG_FILE" -format UDZO -o "$DMG_FILE"
